msgs: extend ExplorationInfo by machine specification
In order for a robot to give a meaningful report, i.e. to report a type
for a particular machine at a certain position it must also know where
the machines are that it has to identify. Therefore extend the
ExplorationSignal message to have a small machine specification
concisting of name and position. Now the robot can properly reference
identified machine types.